feat(upload): enable max files per upload configuration
This is related to CVE-2023-24998 and enables to configure the SWORD
servlet to allow only as much files uploaded in one go as specified.
As we cannot give another sane default than "no limit", a user of the
lib still needs to add a real limit to be on the safe side!